chore(zero-cache): one more script reorg
