+++ packages/zero-cache/src/scripts/deploy-permissions.ts
@@ -1,67 +1,19 @@
 import {consoleLogSink, LogContext} from '@rocicorp/logger';
 import 'dotenv/config';
 import {writeFile} from 'node:fs/promises';
-import {basename, dirname, join, relative, resolve} from 'node:path';
-import {fileURLToPath} from 'node:url';
 import {literal} from 'pg-format';
-import {tsImport} from 'tsx/esm/api';
 import {parseOptions} from '../../../shared/src/options.ts';
-import * as v from '../../../shared/src/valita.ts';
-import {
-  permissionsConfigSchema,
-  type PermissionsConfig,
-} from '../../../zero-schema/src/compiled-permissions.ts';
-import {isSchemaConfig} from '../../../zero-schema/src/schema-config.ts';
+import {type PermissionsConfig} from '../../../zero-schema/src/compiled-permissions.ts';
 import {ZERO_ENV_VAR_PREFIX} from '../config/zero-config.ts';
 import {ensureGlobalTables} from '../services/change-source/pg/schema/shard.ts';
 import {pgClient, type PostgresDB} from '../types/pg.ts';
-import {deployPermissionsOptions} from './deploy-permissions-options.ts';
+import {deployPermissionsOptions, loadPermissions} from './permissions.ts';
 
 const config = parseOptions(
   deployPermissionsOptions,
   process.argv.slice(2),
   ZERO_ENV_VAR_PREFIX,
 );
-
-export async function loadPermissions(
-  lc: LogContext,
-  schema: typeof config.schema,
-): Promise<PermissionsConfig> {
-  lc.info?.(`Loading permissions from ${schema.path}`);
-  const dir = dirname(fileURLToPath(import.meta.url));
-  const absoluteSchemaPath = resolve(config.schema.path);
-  let relativePath = join(
-    relative(dir, dirname(absoluteSchemaPath)),
-    basename(absoluteSchemaPath),
-  );
-
-  // tsImport doesn't expect to receive slashes in the Windows format when running
-  // on Windows. They need to be converted to *nix format.
-  relativePath = relativePath.replace(/\\/g, '/');
-
-  let module;
-  try {
-    module = await tsImport(relativePath, import.meta.url);
-  } catch (e) {
-    lc.error?.(`Failed to load zero schema from ${absoluteSchemaPath}:`, e);
-    process.exit(1);
-  }
-
-  if (!isSchemaConfig(module)) {
-    throw new Error(
-      `Schema file ${schema.path} must export [schema] and [permissions].`,
-    );
-  }
-  try {
-    const schemaConfig = module;
-    const perms =
-      await (schemaConfig.permissions as unknown as Promise<unknown>);
-    return v.parse(perms, permissionsConfigSchema);
-  } catch (e) {
-    lc.error?.(`Failed to parse Permissions object`, e);
-    process.exit(1);
-  }
-}
 
 async function validatePermissions(
   _lc: LogContext,
@@ -119,7 +71,7 @@ async function writePermissionsFile(
 
 const lc = new LogContext('debug', {}, consoleLogSink);
 
-const permissions = await loadPermissions(lc, config.schema);
+const permissions = await loadPermissions(lc, config.schema.path);
 if (config.output.file) {
   await writePermissionsFile(
     lc,
